+++ web/src/components/people_subs.rs
@@ -5,8 +5,8 @@ use crate::components::audio::on_play_pause;
 use crate::components::audio::AudioPlayer;
 use crate::components::context::{AppState, ExpandedDescriptions, UIState};
 use crate::components::gen_funcs::sanitize_html_with_blank_target;
+use crate::requests::episode::Episode;
 use crate::requests::people_req::{self, PersonSubscription};
-use crate::requests::pod_req::Episode;
 use i18nrs::yew::use_translation;
 use yew::prelude::*;
 use yew::{function_component, html, Html};
