+++ web/src/components/virtual_list.rs
@@ -2,11 +2,11 @@ use super::gen_components::{on_shownotes_click, ContextButton, EpisodeModal, Fal
 use super::gen_funcs::{format_datetime, match_date_format, parse_date};
 use crate::components::audio::on_play_pause;
 use crate::components::context::{AppState, UIState};
+use crate::components::episode_list_item::EpisodeListItem;
 use crate::components::gen_funcs::{format_time, strip_images_from_html};
 use crate::components::gen_funcs::{sanitize_html_with_blank_target, truncate_description};
 use crate::components::safehtml::SafeHtml;
-use crate::components::episode_list_item::EpisodeListItem;
-use crate::requests::pod_req::Episode;
+use crate::requests::episode::Episode;
 use gloo::events::EventListener;
 use i18nrs::yew::use_translation;
 use std::rc::Rc;
