+++ web/src/requests/mod.rs
@@ -1,9 +1,9 @@
+pub(crate) mod episode;
 pub(crate) mod login_requests;
 pub(crate) mod models;
 pub(crate) mod people_req;
 pub(crate) mod pod_req;
 pub(crate) mod search_pods;
-pub(crate) mod task_reqs;
-
 pub(crate) mod setting_reqs;
 pub(crate) mod stat_reqs;
+pub(crate) mod task_reqs;
