+++ web/src/requests/pod_req.rs
@@ -2,6 +2,7 @@ use anyhow::{Context, Error};
 // use futures_util::stream::StreamExt;
 use crate::components::context::AppState;
 use crate::components::notification_center::TaskProgress;
+use crate::requests::episode::Episode;
 use futures::StreamExt;
 use gloo::net::websocket::WebSocketError;
 use gloo::net::websocket::{futures::WebSocket, Message};
@@ -14,124 +15,6 @@ use std::fmt;
 use wasm_bindgen::JsCast;
 use web_sys::console;
 use yewdux::Dispatch;
-
-fn bool_from_int<'de, D>(deserializer: D) -> Result<bool, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    struct BoolOrIntVisitor;
-
-    impl<'de> serde::de::Visitor<'de> for BoolOrIntVisitor {
-        type Value = bool;
-
-        fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-            formatter.write_str("a boolean or an integer")
-        }
-
-        fn visit_bool<E>(self, value: bool) -> Result<Self::Value, E> {
-            Ok(value)
-        }
-
-        fn visit_i64<E>(self, value: i64) -> Result<Self::Value, E>
-        where
-            E: serde::de::Error,
-        {
-            Ok(value != 0)
-        }
-
-        fn visit_u64<E>(self, value: u64) -> Result<Self::Value, E>
-        where
-            E: serde::de::Error,
-        {
-            Ok(value != 0)
-        }
-    }
-
-    deserializer.deserialize_any(BoolOrIntVisitor)
-}
-
-fn null_as_zero<'de, D>(deserializer: D) -> Result<i32, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    // Option<i32> will deserialize null as None, or a number as Some(value)
-    let opt = Option::<i32>::deserialize(deserializer)?;
-    Ok(opt.unwrap_or_default())
-}
-
-#[derive(Deserialize, Debug, PartialEq, Clone, Serialize, Default)]
-#[serde(default)]
-#[allow(non_snake_case)]
-pub struct Episode {
-    pub podcastid: i32,
-    pub podcastname: String,
-    #[serde(alias = "Episodetitle")]
-    pub episodetitle: String,
-    pub description: String,
-    pub artworkurl: String,
-    pub author: String,
-    pub categories: Option<HashMap<String, String>>,
-    #[serde(alias = "Episodedescription")]
-    pub episodedescription: String,
-    pub episodecount: Option<i32>,
-    pub feedurl: String,
-    pub websiteurl: String,
-    pub explicit: i32,
-    pub userid: i32,
-    #[serde(alias = "Episodeid")]
-    pub episodeid: i32,
-    #[serde(alias = "Episodeurl")]
-    pub episodeurl: String,
-    #[serde(alias = "Episodeartwork")]
-    pub episodeartwork: String,
-    #[serde(alias = "Episodepubdate")]
-    pub episodepubdate: String,
-    #[serde(alias = "Episodeduration")]
-    pub episodeduration: i32,
-    #[serde(alias = "Listenduration", deserialize_with = "null_as_zero")]
-    pub listenduration: i32,
-    #[serde(alias = "Completed")]
-    pub completed: bool,
-    pub saved: bool,
-    pub queued: bool,
-    pub downloaded: bool,
-    pub is_youtube: bool,
-    pub guid: String,
-    pub queueposition: Option<i32>,
-    pub downloadedlocation: Option<String>,
-}
-
-impl Episode {
-    pub fn get_episode_artwork(&self) -> String {
-        self.episodeartwork.clone()
-    }
-
-    pub fn get_episode_title(&self) -> String {
-        self.episodetitle.clone()
-    }
-
-    pub fn get_is_youtube(&self) -> bool {
-        self.is_youtube
-    }
-
-    pub fn get_episode_id(&self, _fallback_id: Option<i32>) -> i32 {
-        self.episodeid.clone()
-    }
-
-    pub fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    /// Parse json Value to populate an Episode struct
-    /// Any field not included in the json will be filled with its default value
-    ///
-    /// Surely there is a better way? Just converts the Value back to string, then deserializes
-    /// that string into an Espiode.
-    pub fn from_json(json: &serde_json::value::Value) -> Result<Self, serde_json::Error> {
-        let j = serde_json::to_string(json)?;
-        serde_json::from_str::<Self>(&j)
-    }
-}
 
 #[derive(Deserialize, Debug, PartialEq, Clone)]
 pub struct RecentEps {
@@ -428,6 +311,41 @@ pub async fn call_remove_podcasts_name(
             response.status_text()
         )))
     }
+}
+
+fn bool_from_int<'de, D>(deserializer: D) -> Result<bool, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    struct BoolOrIntVisitor;
+
+    impl<'de> serde::de::Visitor<'de> for BoolOrIntVisitor {
+        type Value = bool;
+
+        fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
+            formatter.write_str("a boolean or an integer")
+        }
+
+        fn visit_bool<E>(self, value: bool) -> Result<Self::Value, E> {
+            Ok(value)
+        }
+
+        fn visit_i64<E>(self, value: i64) -> Result<Self::Value, E>
+        where
+            E: serde::de::Error,
+        {
+            Ok(value != 0)
+        }
+
+        fn visit_u64<E>(self, value: u64) -> Result<Self::Value, E>
+        where
+            E: serde::de::Error,
+        {
+            Ok(value != 0)
+        }
+    }
+
+    deserializer.deserialize_any(BoolOrIntVisitor)
 }
 
 #[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
@@ -3551,9 +3469,12 @@ pub async fn call_update_episode_duration(
     request_data: &UpdateEpisodeDurationRequest,
 ) -> Result<String, Error> {
     let url = format!("{}/api/data/update_episode_duration", server_name);
-    let api_key_ref = api_key.as_deref().ok_or_else(|| anyhow::Error::msg("API key is missing"))?;
+    let api_key_ref = api_key
+        .as_deref()
+        .ok_or_else(|| anyhow::Error::msg("API key is missing"))?;
 
-    let request_body = serde_json::to_string(request_data).map_err(|e| anyhow::Error::msg(format!("Serialization Error: {}", e)))?;
+    let request_body = serde_json::to_string(request_data)
+        .map_err(|e| anyhow::Error::msg(format!("Serialization Error: {}", e)))?;
 
     let response = Request::post(&url)
         .header("Api-Key", api_key_ref)
@@ -3563,10 +3484,14 @@ pub async fn call_update_episode_duration(
         .await?;
 
     if response.ok() {
-        let response_body: UpdateEpisodeDurationResponse = response.json().await.map_err(|e| anyhow::Error::new(e))?;
+        let response_body: UpdateEpisodeDurationResponse =
+            response.json().await.map_err(|e| anyhow::Error::new(e))?;
         Ok(response_body.detail)
     } else {
-        let error_text = response.text().await.unwrap_or_else(|_| String::from("Failed to read error message"));
+        let error_text = response
+            .text()
+            .await
+            .unwrap_or_else(|_| String::from("Failed to read error message"));
         Err(anyhow::Error::msg(format!(
             "Failed to update episode duration: {} - {}",
             response.status_text(),
